tests/wpt/web-platform-tests/notifications/event-onshow.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Web Notifications Test: notification - onshow (basic)</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="author" title="Xin Liu" href="mailto:xinx.liu@intel.com">
<link rel="help" title="5 API" href="http://www.w3.org/TR/notifications/">
<meta name="flags" content="">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
  <div id="log"></div>
  <script>
    t = async_test("the notification.onshow event can be invoked", {timeout: 50000});
    var notification = new Notification("New Email Received", {
      body: "Room 101"
    });
    notification.onshow = t.step_func(function(e) {
      assert_equals(Object.prototype.toString.call(e), "[object Event]", "the type of event");
      t.done();
    });
  </script>
</body>
